requirements.txt
absl-py==1.4.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncssh==2.14.2
atpublic==4.0
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
billiard==4.2.0
black==23.12.0
bleach==6.1.0
blinker==1.6.2
cachetools==5.3.1
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
configobj==5.0.8
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.8
distro==1.8.0
dm-tree==0.1.8
docker==6.1.3
dpath==2.1.6
dulwich==0.21.7
dvc==3.35.0
dvc-data==3.1.0
dvc-http==2.32.0
dvc-objects==3.0.0
dvc-render==1.0.0
dvc-studio-client==0.18.0
dvc-task==0.3.0
ensure==1.0.4
entrypoints==0.4
etils==1.5.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.9.1
funcy==2.0
gast==0.4.0
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grandalf==0.8
greenlet==3.0.2
grpcio==1.58.0
gto==1.6.1
gunicorn==21.2.0
h5py==3.9.0
huggingface-hub==0.20.2
hydra-core==1.3.2
identify==2.5.33
idna==3.4
importlib-metadata==7.0.0
importlib-resources==6.1.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.13.1
kiwisolver==1.4.5
kombu==5.3.4
libclang==16.0.6
Mako==1.3.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.11
mistune==3.0.2
ml-dtypes==0.2.0
mlflow==2.9.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.10
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.0.1
platformdirs==3.11.0
pre-commit==3.6.0
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
pygit2==1.13.3
Pygments==2.17.2
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-box==7.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.4
scmrepo==2.0.2
seaborn==0.13.0
semver==3.0.2
Send2Trash==1.8.2
shortuuid==1.0.11
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
sqltrie==0.10.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
tomlkit==0.12.3
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.37.1
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.17
vine==5.1.0
virtualenv==20.25.0
voluptuous==0.14.1
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
